Refactor: Next.js AppRouter 구조에 맞도록 components, constants, hooks, utils 폴더 위치 조정
