+++ src/constants/common/BottomTabNav.ts
